Simplify my life by removing paranoia tests
Commonmarker currently has three different test suites related to the
gem build:

1. The compilation output is tested
2. The gem can be installed on another platform
3. A generic version of the gem can be installed on a platform with
sufficient Ruby/Rust tooling

Of these, only the first is the most relevant. 2 suggests an error in
the build pipeline, and 3 opens the door to supporting platforms that
are uncommon. In order to minimize support work
on my end, I'm going to nix the last two tests; as
well, I can abstract the first test into a GitHub Action for other
oxidizers to use.