Release v2.0.0
This is a major release in the sense that there's a breaking API change,
but not in the sense that anything earth-shatteringly urgent has
occurred.

In certain situations, user-provided options were not being merged
correctly. This meant that default config values were being ignored
erroneously.
This was a bug, which has been fixed. Unfortunately, it's possible that
downstream projects were relying on the old, broken behavior, so a
simple patch release wouldn't do.

If you're _not_ passing any options to `to_html` and everything works
fine for you, than this should be an upgrade without side-effects.

If you _are_ passing options to `to_html` (i.e. overridding the
defaults), then consult the README for a list of options and
their default values. Make sure that the default values given to
`to_html` is acceptable to your project!